working Auction.sol
I am not sure how to show multiple commits because I didnt get remixd to work - sorry ;( 

You will need to copy and paste this in solidity to get make it run. 
the auctions includes a timer where you can set the time of the auction in seconds. There is also a function that will allow you to see the time remaining 
the auction also includes a game. every time a person places a bid a dice is rolled and if the bidder does roll a one the bid will go through. However if the bidder roles a one then they will get a message saying they can't place a bid and will have to run the method again
The auction also contains a comment feature.